gtk3: Don't depend on a11y to detect tooltip win
In RunDialog::windowOpened, don't rely on the
the VCLXWindow currently usually is also being
the XAccessible for a vcl::Window, which is an
implementation detail and subject to change, see
commit messsage of

        Change-Id: Id81ab5f90955ecc600e179164b5f9c7a771182d1
        Author: Michael Weghorn <m.weghorn@posteo.de>
        Date:   Wed May 21 11:03:56 2025 +0200

            sw a11y test: Don't rely on XWindow being XAccessible

for further background.

Instead, retrieve the vcl::Window and check the window
type of that one directly to detect a tooltip.
(See the mapping from window type to accessible role
in Window::getDefaultAccessibleRole.)

If continuing to use the a11y layer were necessary for
any reason, an alternative could be to get the
vcl::Window's XAccessible without it relying on it
being the VCLXWindow, see the above-mentioned commit
for an example.

Sample scenario that triggers this code path and
still behaves the same as without this commit in
place:

* start Writer with the gtk3 VCL plugin on Linux, create
  new doc
* press Ctrl+N to create another new document
* press Ctrl+S to open the file dialog for the second doc
* switch back to the first doc
* press Ctrl+N to open yet another doc

-> The file dialog is popped down after a few moments
because the logic in RunDialog::windowOpened still
detects that the new window isn't a tooltip.

Change-Id: I66c9d4c87c7ba96502d08ebc8ec52fdaabe132ba
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185603
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins