tdf#164970 add a checkbox for enabling/disabling application theming
the "enable application theming" checkbox allows the users to
enable/disable ui theming. in the beginning, theming was enabled by
default but then after [quite some discussion], it was disabled. so if
the user wanted to enable it again, he had to dig through the expert
settings to set an integer config value to 1. this checkbox makes it
convenient.

[quite some discussion]: https://gerrit.libreoffice.org/c/core/+/183490

Change-Id: I5c4228d103d41b8739ff0b72743f962bb20bdb50
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185605
Reviewed-by: Sahil Gautam <sahil.gautam.extern@allotropia.de>
Tested-by: Jenkins