tdf#164970 add checkbox for toggling document color customization
after themes was released in 25.2, there were quite a few bug reports on
bugszilla like [broken dark mode], and some on social platforms like one
[bug reported on X]. this also [broke old presentations], and in that
case it was more problematic than writer or calc as in case of impress
it is assumed that "what you see is what you get".

this happened because the documents didn't have the backgroung color
attribute set and the document canvas color is dark in dark mode. the
solution for this is to not override the document canvas colors by
default and provide users with a choice to enable it if they want.

[broken dark mode]: https://bugs.documentfoundation.org/show_bug.cgi?id=164838
[bug reported on X]: https://x.com/pritamckn25/status/1915068541801210201
[broke old presenatations]: https://bugs.documentfoundation.org/show_bug.cgi?id=165803

Change-Id: I613b5f0b6e4e1f8f2b8f9fddb50f98286dbdc481
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185388
Tested-by: Jenkins
Reviewed-by: Sahil Gautam <sahil.gautam.extern@allotropia.de>