toolkit a11y: Don't rely on control peer being XAccessible
When getting the accessible context for a
UnoControl in non-design mode, don't rely/depend
on the implementation detail that the VCLXWindow
implements the XAccessible interface for its
vcl::Window, as this is subject to change.

Instead, retrieve the vcl::Window and then
get its accessible (context).

See also commit messsage of

    Change-Id: Id81ab5f90955ecc600e179164b5f9c7a771182d1
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed May 21 11:03:56 2025 +0200

        sw a11y test: Don't rely on XWindow being XAccessible

for more background.

Sample scenario to trigger this code path:

* start Orca screen reader on Linux
* start Writer with gtk3 VCL plugin
* enable "Form Controls" toolbar in "View" -> "Toolbars"
* insert a Check Box (or radio button) control
* disable design mode by clicking the corresponding button
  in the form controls toolbar
* if not already the case: select the control

At the moment, the accessible context is exactly the
same with or without this commit in place, but
the previous approach relying on VCLXWindow being
the vcl::Window's XAccessible would break if that
no longer is the case in the future.

Change-Id: I9aa697e1d7a25a10ccb5469f5dd2a84a4f00fbfc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185617
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>