tdf#166319 sw interdependent redlines: handle accept of ins-then-fmt's fmt
The bugdoc has <ins>AA<format>BB</format>CC</ins> in it, accepting the
"BB" part resulted in a fully accepted portion, while the expected
result is to accept the underlying insert, like in Word.

The problem is that the accept code is not aware that the format has an
underlying insert, so the format gets accepted and as a side effect the
insert redline is gone, too.

Fix this by extending DocumentRedlineManager::AcceptRedlineRange() to
just get rid of the inner redline in the insert-then-format case, which
results in a plain format redline, as wanted. lcl_DeleteInnerRedline()
can do exactly this, added already in commit
f468d61755fad1e38e481f8f439bbce9161a3aa7 (tdf#166319 sw interdependent
redlines: handle reject of delete under format, 2025-05-16).

Connecting this (effectively) insert redline (insert-then-format, to be
precise) with the surrounding plain insert redlines is not done yet,
though.

Change-Id: I64e2ab1f6a6ce7872c3e4ea7b0bfb0c542c611ed
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185653
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins