wina11y: Don't rely on win accessible being VCLXWindow
While at the moment, the XAccessible for many vcl::Window
subclasses is a VCLXWindow, this is an implementation
detail that is subject to change.
Switch away from relying on that implementation detail,
and instead rely on the XAccessibleContext being a
VCLXAccessibleComponent, which will remain true even
if that one will be created without VCLXWindow being
involved in the future.

See also

	commit b1f3e41f1015f0166dca1382c5fa1d1223f72fb8
	Author: Michael Weghorn <m.weghorn@posteo.de>
	Date:   Wed May 21 11:03:56 2025 +0200

	    sw a11y test: Don't rely on XWindow being XAccessible

for more background (which ports away from an assumption
"the other way around").

Change-Id: Ia2e2fe9d88c05afcde2c8a6d31c20c60cfbd5c49
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185645
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>