vcl a11y: Let PushButton create its accessible directly
So far, the VCLXButton was the XAccessible implementation
for PushButton (and subclasses), which got returned in
Window::CreateAccessible.

VCLXButton's implementation of XAccessible::getAccessibleContext
was then responsible for creating the XAccessibleContext,
a VCLXAccessibleButton.

Drop that separation between XAccessible and XAccessibleContext,
let VCLXAccessibleButton also implement the XAccessible interface
itself, and override Window::CreateAccessible in
PushButton to create the VCLXAccessibleButton directly without
requiring the VCLXButton and AccessibleFactory for this.

This is one step towards unifying the approach for all vcl::Window
subclasses (some were already creating their accessibles
directly), makes responsibilities clearer (currently
shared between vcl::Window and VCLXWindow) and prepares
for more upcoming a11y changes (towards making
VCLXAccessibleComponent/OAccessibleComponentHelper implement both,
XAccessible and XAccessibleContext).

Similar changes for other vcl::Window subclasses will follow.

Internal code previously depending on the fact that a vcl::Window's
VCLXWindow/XWindow was its XAccessible was adjusted previously in
commits like

    commit b1f3e41f1015f0166dca1382c5fa1d1223f72fb8
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed May 21 11:03:56 2025 +0200

        sw a11y test: Don't rely on XWindow being XAccessible

If this commit causes any issues, there might be
another place requiring a similar change.

According to the discussion in [1], there might be third-party
code (unrelated to a11y except for using UNO a11y API) also
relying on the assumption that the vcl::Window's XAccessible
is its VCLXWindow, but the UNO a11y API is an internal
implementation detail explicitly unpublished in

    commit 70626249cd247d9acdad417b8eaf252bae22c059
    Date:   Thu Nov 29 00:27:03 2012 +0100

        API CHANGE a11y unpublishing and add/removeListener rename.

and has no stability guarantuee whatsoever and is therefore
not meant to be used by external users. If any third-party
code was relying on this implementation detail, it should be
ported away from relying on UNO a11y API.

If there were a desperate need to still be able to
retrieve the XAccessibleContext via the VCLXWindow/XWindow
(at least temporarily) somehow, a (temporary) workaround
once this commit and corresponding upcoming commits are in place
could be to let VCLXWindow still implement the XAccessible
interface and return the XAccessibleContext from the window's
actual XAccessible in VLCXWindow::getAccessibleContext.
But in my opinion, code relying on that assumption should really
be adjusted instead, also to avoid breaking again when unpublished
UNO a11y API changes again in the future.

[1] https://gerrit.libreoffice.org/c/core/+/185597/comments/48aae563_7c620891

Change-Id: If6fa82008f3347c4f95538ab181eaa50c7d9636b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185670
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins