tdf#164970 [API CHANGE] remove all app background related api/ui/code
previously all this was done in a single [patch] and it made things
quite complicated for me (to change) and for the reviewers as many
things were moving in a single patch.

API CHANGE:
only application background is bitmap customizable and we stored
the user settings for bitmap file name and draw type as properties
of AppBackground color entry in a ColorScheme. That unneccessarily
complicated things as we had to check for the "/BackgroundType" while
Loading and Committing (saving) the colorschemes.

this patch removes the bitmap properties from AppBackground entry
of the colorscheme schema along with all the feature implementation
code. most of the implementation code is added back in a follow up
patch with a new underlying api and ui to make it independent of color
customization.

[patch]: https://gerrit.libreoffice.org/c/core/+/184544

Change-Id: I71fee27467be861b610934c213cf80fbae037327
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185697
Reviewed-by: Sahil Gautam <sahil.gautam.extern@allotropia.de>
Tested-by: Jenkins