tdf#37153 tdf#165478 sw: add mso-compat flag doNotVertAlignCellWithSp
This patch depends on
    NFC tdf#37153 ww8import: replace 'unknown' with compatibility names

The problem was that the cell vertical orientation should not
ALWAYS be forced to top when the cell contains an floating shape.

This replaces my 6.0 commit 7a9fb40cb07de8c2ea33f92735be5008d30d6704
which incorrectly applied to all MSO documents. All DOC files - yes,
but it should only affect DOCX with compat flag doNotVertAlignCellWithSp
(Don't vertically align table cells with shapes).

It appears that DOC format should ALWAYS be considered
as having this flag enabled, even if it is set to OFF
(and it IS set to off somehow in the unit tests).
At least that is how MS Word 2010 seems to deal with it,
and MS Word 2003 (which doesn't detail these layout flags)
also works the same way (even for the DOCX - which reinforces all DOC).

- open the unit test tdf37153_considerWrapOnObjPos.doc
--   see the image/paragraph marker at the top of the cell.
- delete the image
--   see the paragraph marker drop to the bottom of the cell.

make CppunitTest_sw_ww8export2 CPPUNIT_TEST_NAME=testTdf37153

make CppunitTest_sw_ooxmlexport25 \
    CPPUNIT_TEST_NAME=testTdf165478_bottomAligned

For ooxmlexport24's tdf37153_considerWrapOnObjPos.docx,
I probably created that unit test after testing in Word 2003.
There it behaved identically to the DOC version,
and that becomes no surprise when I now find out that
doNotVertAlignCellWithSp is a flag introduced for Word 2007.
Since it was testing for an ancient layout - I changed it.

Change-Id: Id1ef078a19605532bbb694532cbe7f3cf82fb565
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185695
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins