tdf#149207 fix image not getting a unique name
Using "background" as the name, CheckNamedItem doesn't correctly
determine that another XFillBitmapItem exists with an same
"background" name already, which is then a problem when exporting
the document and from all "background" images only one will be
exported.

regression from
  commit ec7ba61a6164c805f5a71b077715b7e1521a2d62
  Author: Noel Grandin <noel.grandin@collabora.co.uk>
  Date:   Wed Apr 17 15:19:25 2019 +0200
  simplify SfxPoolItemArray_Impl (tdf#81765 related)

I am not sure why the above commit caused this problem.

Right now, the reason CheckNamedItem doesn't work is because
with other subsequent changes to the pool infrastructure, the
new PoolItem is already in the surrogates/NameOrIndex array
and so it finds itself when scanning the array.

But back then, the new PoolItem should not have been inserted
yet.

Any way, fix it by making CheckNamedItem exclude itself when
scanning the surrogates data.

Added a test for uniqueness of names.

Change-Id: I3c04065a4875a21649ad8141243cc39a8d1a5b3c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185719
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>