tdf#166319 sw interdependent redlines: combine on accept of ins-then-fmt's fmt
The bugdoc has <ins>AA<format>BB</format>CC</ins> in it, accepting the
"BB" part resulted in a format-only redline in the middle (good), but
"AA" and "CC" wasn't accepted, as expected.

The first trouble was that SwRedlineTable::getConnectedArea() calling
lcl_CanCombineWithRange() considered combining AA with BB when the
cursor was at AA, but it didn't consider combining BB with AA when the
curor was at BB. There doesn't seem to be a reason why combining one
type with the underlying type of a next redline should be only done in
one direction, so allow both directions. With this, at least
AcceptRedlineRange() gets the right redline indexes, so it considers the
two surrounding redlines.

The other trouble was that DocumentRedlineManager::AcceptRedlineRange()
only tried to combine a redline with 1 type together with an other
redline with 2 types (assuming that perhaps the 2nd/underlying type is a
match), and didn't try to combine a redline with 2 types with an other
redline with 1 type. Allow that case at least for the insert vs
insert-then-format case when the cursor is at the insert-then-format
redline.

The reject of insert-then-format's format is still broken, though.

Change-Id: Ic8ab03e9cd2c24506cf98c5e82c09464eab8f6b5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185701
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>