tdf#162826: ipdf: Convert for encrypted, correctly versioned, files
convertToHighestSupported downconverts PDFs to a magical 1.6 version
due to some historic bug; it's also used in a corner of the ipdf
vector convert to force a conversion on a buggy input file.

Use it to also force a conversion in the encrypted case even if the
input was of the expected version.
Pass up a flag to indicate if the file was encrypted.

Change-Id: I3a8d17851349157ec5b8e7d9a2548c1a21a198a7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185757
Reviewed-by: David Gilbert <freedesktop@treblig.org>
Tested-by: Jenkins