tdf#165094: do not skip empty page because of wrong page descriptor
The problem was, that the layout process created three pages initially,
with automatic page break before paragraph 26 (that is controlled by
SwLayHelper::mnMaxParaPerPage and checked in SwLayHelper::CheckInsert),
and manual page break before the hidden paragraph; then the paragraphs
26-45 moved back to page 1; and then the content of page 3 attempted
to move back. Looking for a page to move to (in SwFlowFrame::MoveBwd),
it found page 2 as the candidate target; but in SwFrame::WrongPageDesc,
it concluded that the page 2 can't be used (because of different page
style, or a difference in even/odd/first format). Thus, that page was
skipped, the content moved to page 1, effectively ignoring the break.

The check in SwFrame::WrongPageDesc, called from SwFlowFrm::MoveBwd,
was added in commit 84a3db80b4fd66c6854b3135b5f69b61fd828e62 (initial
import, 2000-09-18). Simultaneously, bCheckPageDescs was set there,
and the page formats were checked eventually. I think, that skipping
pages because of wrong page formats is unneeded because of that final
check; this patch removes the problrmatic conditions.

(cherry picked from commit 67308193f6ee8a21b20043fbd3e831a8c455f40a)

Change-Id: Ife0b5591c68adb46f7c8fdde0e852e9e326dc38a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185751
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-on: https://gerrit.collaboraoffice.com/c/core/+/4786
Tested-by: Jenkins CPCI <releng@collaboraoffice.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>