don't probe meson with "python meson runpython --version"
runpython is internal flag that's not supported in all versions, and was
used to prevent testing "python --version" in case MESON was empty, but
that's now checked for explicitly, so that can be removed.

Also fix logic error when selecting pyuno python version / the one that
has the final say on what python will be used for the build. Skipping
the version bump to 3.7 or later when PYTHON_FOR_BUILD was already set
just nullifies the earlier check wrt harfbuzz/whether meson is needed.

Change-Id: I50bb74a806f8c48d91e12616b69e64c2caa4e7be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185754
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Tested-by: Jenkins