tdf#162826: vcl: Make readPDF stash unencrypted pdf
On an insert of a PDF document which is encrypted, make sure
that the unencrypted version is the one we save in the combined
LO document.   The user can always add a password on that.
There's no sane way to ask for, potentially multiple, embedded
PDFs, and since we have the bitmap anyway there's no point.

Change-Id: I06f3a406f8664a9fba21a1b26a3bc35e7492b37b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185759
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: David Gilbert <freedesktop@treblig.org>