sd: write embedded fonts with a more descriptive name
If MSO writes the name of fonts with a generic name like "font1",
this doesn't mean that we need to do the same thing. Use a better
more descriptive names of the fonts that include the font family
name and the font type (regular, bold, italic,...).

Test assert needed to be removed - much harder to test this now,
but not really that critical.

Change-Id: I96b9f6ee9d34fa7a030fdf0496ef4078bd3dc946
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185763
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit ff4ff571defc5e62c469799a05a562012fc40ec7)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186168
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>