sd: better way to check if a font is used for PPTX font embedding
Check all the objects and style that is accosiated with the object
if the font is used. This way we don't include fonts that are not
actually used in the document. The isUsed on the document is not
percise enough in this case.

Change-Id: If11d80ffe04b28688698babb560ac0be84bbed12
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185764
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
(cherry picked from commit e888a6d9eb2af220a9c8c0a3483e37d1f53f628f)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186223
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>