[API CHANGE] uitest: Add API to get XUIObject for specific XWindow
Add a XUITest::getWindow method that allows to get the
XUIObject for a given XWindow.

The XUITest interface is internal/unpublished API used
for UI tests.

The newly introduced method will be used in an upcoming
commit in sw/qa/uitest/table/tdf146145.py to port
that test away from depending on a11y implementation
details that are about to change.

Change-Id: I6fcbee3ce894ad82ec42fa034b85544b50408779
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185773
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>