sw uitest: Port tdf146145 away from using a11y API
Stop using the UNO a11y API in that UI test in
order to check whether the "Accept" toolbar item
in the Track Changes toolbar is enabled or not.

The previous logic was depending on the implementation
detail that ToolBox's XWindow is also its XAccessible,
which will no longer be the case after upcoming commit

    Change-Id: Ib6a6746f7fcbcacddc4cc05a00cf63fe04407f80
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 26 09:08:41 2025 +0200

        vcl a11y: Let ToolBox create its accessible directly

Instead, use the API newly introduced in previous commit

    Change-Id: I6fcbee3ce894ad82ec42fa034b85544b50408779
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 26 08:26:37 2025 +0200

        [API CHANGE] uitest: Add API to get XUIObject for specific XWindow

to get a UI test object for the toolbar.
For checking the state of the toolbar item, check the
toolbar child infos, as the toolbar items and their state
are provided that way since

    Change-Id: I6064ad3481e1cef694762c321e244097d8c3029b
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 26 08:57:32 2025 +0200

        vcl uitest: Add support for toolbar items

This also allows to use the toolbar item's text
to identify the one of interest, instead of having
to rely on the toolbar item being at a particular index.

To run the test:

    make UITest_sw_table UITEST_TEST_NAME="tdf146145.tdf146145.test_tdf146145"

Change-Id: Ice15d7a9e48c27a257bd6943f473b75c2526e3f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185775
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins