vcl uitest: Add support for toolbar items
Add new class ToolBoxItemUIObject that provides
access to information about ToolBox (toolbar)
items in UI tests.
Initially, report the item's text and enabled
state. (This can be extended as needed if there
is a need in the future.)

Extend ToolBoxUIObject to report the ToolBox
items as children in addition to those children
retrieved via the WindowUIObject base class logic.

This will be used in an upcoming commit in
sw/qa/uitest/table/tdf146145.py to port
that test away from depending on a11y implementation
details that are about to change.

Change-Id: I6064ad3481e1cef694762c321e244097d8c3029b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185774
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>