vcl a11y: Set DateField/TimeField window type in ctor
So far, a DateField (or subclass) would be created
with WindowType::SPINFIELD due to its SpinField
base class (and not passing any explicit type to
the base class ctor).

However, when an accessible context would later be
retrieved, VCLXDateField::CreateAccessibleContext
would set WindowType::DATEFIELD on the control
before creating an XAccessibleContext via the
AccessibleFactory.

Changing the window type later that way seems odd,
and implies that the window type would be different
depending on whether or not accessibility is active.

Change that to set the DATEFIELD window type in the
ctor already.

The same applies for the TimeField and
WindowType::TIMEFIELD.

Should that have any unwanted side-effect, then
it might be that some code path currently only handling
the WindowType::SPINFIELD type might need to be
adjusted to handle WindowType::DATEFIELD and/or
WindowType::TIMEFIELD the same way.

Change-Id: Ifb7452de3e011a73aa798d1ea0302bbbc74ed7d2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185837
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>