vcl a11y: Let vcl::Window create accessible directly, drop a11y factory
This is basically the vcl::Window base class equivalent of

    commit f82aac71ede9c3f2dc059ca4c8b24f60e0841933
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu May 22 12:30:12 2025 +0200

        vcl a11y: Let PushButton create its accessible directly

, see that commit's message for more background and tips on
what to do if this commit causes any issues.

Now that Window::CreateAccessible and all overrides in
subclasses no longer make use of the VCLXWindow for
the XAccessible, VCLXWindow::CreateAccessibleContext
should no longer be called. Add an assert that would
trigger if that method was still called somehow.

Drop the now unused AccessibleFactory.

Further cleanup will be done in separate commits.

Java a11y tests that were previously relying on the
implementation detail that the XAccessible for many
vcl::Window subclasses was their VCLXWindow have been
adjusted in previous commit

    Change-Id: Id96a2104d87c998b0bdcbce36451499484d81272
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Jun 18 17:49:48 2025 +0200

        java a11y tests: Don't rely on XWindow being XAccessible

Change-Id: I25de1ab0f8b47b12346abd01bad05cb54fb75f52
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185843
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>