tdf#147299 vcl: Set Microsoft ja IME to vertical mode in vertical text
Fixes a Windows IME integration bug causing the Microsoft Japanese IME
to display inappropriately rotated text or overlap the composition
string when the text direction is vertical.

Windows IMEs do not expose an explicit option to enable vertical
writing. Instead, this mode is set by providing a font face with a
leading '@' sign. This behavior is not documented, but is long-standing
and is the standard convention on Windows in other contexts.

Change-Id: I9ebe45cf2c48b5f175d369daaa891f86fdc77c27
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185870
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>