tdf#166319 sw interdependent redlines: handle reject of ins-then-fmt's fmt
The bugdoc has <ins>AA<format>BB</format>CC</ins> in it, rejecting the
BB part resulted in <ins>AA</ins>BB<ins>CC</ins>, while the expected
result would be to reject the underlying insert of BB.

The problem is that the reject code is not aware that the format has an
underlying insert, so the format gets rejected, and as a side effect the
insert redline is gone, too.

Fix this by extending sw::DocumentRedlineManager::RejectRedlineRange()
to recognize this insert-then-format case: act like rejecting the
insert, not like rejecting the format.

The surrounding plain inserts are not yet rejected, though. Also update
extend sw doc model xml dump code that was useful to see while working
on this.

Change-Id: I6718b9c306239837392ae7cd3f0d28aef6b4f71e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185891
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins