tdf#130857 cui: Use separate columns in toolbar tab page
Use a separate GtkTreeViewColumn for each of the
cell renderers in the "View" -> "User Interface"
dialog's new "Toolbars" tab page introduced in

    commit 4ab26f4ea0c627844531144a6a903c357a604a5e
    Date:   Mon May 26 12:03:40 2025 +0200

        Resolves tdf#158880 - Let users select multiple toolbars at once

The weld::TreeView API also uses separate columns,
and the Qt implementation in QtInstanceTreeView relies
on those also being separate columns in the .ui file.

This prepares for declaring support for that tab
page in QtInstanceBuilder in an upcoming commit.

Change-Id: I2f73518e17a34f22b79b563af7ec818692f3edb6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185882
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>