tdf#130857 qt weld: Support new toolbar tab page
Declare support for the "View" -> "User Interface"
dialog's new "Toolbars" tab page introduced in

    commit 4ab26f4ea0c627844531144a6a903c357a604a5e
    Date:   Mon May 26 12:03:40 2025 +0200

        Resolves tdf#158880 - Let users select multiple toolbars at once

in QtInstanceBuilder. This makes the content of
that tab page show up when using the qt6 VCL
plugin with SAL_VCL_QT_USE_WELDED_WIDGETS=1.

The toolbar visibility does not yet get toggled when
(un)checking the corresponding checkbox, but that
will be implemented in an upcoming commit.

Change-Id: I3d0f4ec98c551a11c210f02800ef6d940aba96fe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185884
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins