tdf#166680 DOCX import: handle nested <w:r> runs
Regression from commit 57450afb768c085df0ba2344aa94b5f843060178 (DOCX
import: declare wps as a supported feature, 2013-12-03), the bugdoc has
a shape that has invalid (but well-formed) <w:r> elements inside a <w:r>
element and Word imports this fine.

We used to be compatible, because <mc:AlternateContent> has
<mc:Choice> and <mc:Fallback> branches and the fallback didn't have
these nested runs and we used to read the fallback.

Fix the problem by extending the generated tokenizer to also accept
<w:r> inside <w:r>.

Note that while this is invalid, not only old Writer but also Word
accepts this.

Change-Id: I52e573f921e95dcc6e2b76b2d46bf2b3a03a88ac
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185972
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins