tdf#166321 Find bar: blank square with gen environment
Regression from
    commit e0d4d178caff1414a9a21fa57f06bc8d4d2c389a
    Author: Noel Grandin <noel.grandin@collabora.co.uk>
    Date:   Mon Jan 13 15:03:05 2025 +0200
    Change alpha behavour of OutputDevice::SetFillColor

Which exposed another place where COL_AUTO is troublesome.

Previouslu, calling set_background(COL_AUTO) used to work
because COL_AUTO has the same value as COL_TRANSPARENT,
and calling
    OutputDevice::setBackground(COL_TRANSPARENT)
would have the same effect as
    OutputDevice::setBackground()

But that is no longer the case, especially when we are drawing to
an OutputDevice which has no alpha layer.
In that case, the alpha value is chopped off the Color object, and
we are left with what is effectively COL_WHITE.

So add some new weld API to make resetting the background to default
less error-prone.

Change-Id: I29987c4c343af40447ed847a617d6b0e9d673f48
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185968
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>