tdf#130857 qt weld: Fix qtToVclStringWithAccelerator for empty string
Return an empty string in qtToVclStringWithAccelerator
when the input string is also an empty string.

The previous logic resulted in "~" being returned because
the initial index of 0 would never change.

This resulted in the

    if(!m_xSearchComponent1PB->get_label().isEmpty() && bSearchComponent1 )

logic in SvxSearchDialog::Construct_Impl to return true
and Buttons without text showing up in the Search and Replace
dialog in a WIP branch adding support for that one for
qt6 and SAL_VCL_QT_USE_WELDED_WIDGETS=1.

Change-Id: Ie24c091676a36179471b8bd863d60838632de83b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185986
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins