tdf#166773: avoid using HighlightTextColor on treeview standard image
... when the image is not on a highlight background.

The problem was, that the non-native-drawn symbol of a highlighted
element was drawn using "highlighted text" color (usually white),
but that only makes sense when the background of such an element
(text, symbol) is itself "highlighted" (dark). In tree view, the
usual rendering only gighlights the text of active element, not
the area of the "expand" symbol. That resulted in white-on-white
rendering.

This change tracks, if the background painted for the element's
items intersects with the area of the symbol, and only uses the
"highlighted text" color, if it does.

Change-Id: I781b3341466419c1e654d7b46c8cc8db9ead7120
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186089
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>